Extract pair/candle type from types.ts
