+++ src/types/index.ts
@@ -4,6 +4,7 @@ export * from './backtest';
 export * from './balance';
 export * from './blacklist';
 export * from './botComparison';
+export * from './candleTypes';
 export * from './chart';
 export * from './daily';
 export * from './exchange';
@@ -12,6 +13,6 @@ export * from './locks';
 export * from './pairlists';
 export * from './plot';
 export * from './profit';
-export * from './tradeStats';
 export * from './trades';
+export * from './tradeStats';
 export * from './types';
